# docs/swagger.yaml
basePath: /
definitions:
  bookmark.createBookmarkInput:
    properties:
      description:
        example: A bookmark description
        maxLength: 255
        type: string
      url:
        example: http://google.com
        maxLength: 2048
        type: string
    required:
    - url
    type: object
  bookmark.updateBookmarkBody:
    properties:
      description:
        example: Google
        maxLength: 255
        type: string
      url:
        example: https://www.google.com
        maxLength: 2048
        type: string
    required:
    - url
    type: object
  link.shortenInputBody:
    properties:
      exp:
        minimum: 300
        type: integer
      url:
        type: string
    required:
    - exp
    - url
    type: object
  model.Bookmark:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
  model.HealthCheckResponse:
    properties:
      instance_id:
        type: string
      message:
        type: string
      service_name:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  response.Message:
    properties:
      details: {}
      message:
        type: string
    type: object
  user.getSelfInfoResponse:
    properties:
      data:
        $ref: '#/definitions/model.User'
    type: object
  user.loginInputBody:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  user.loginResponse:
    properties:
      data:
        type: string
      message:
        type: string
    type: object
  user.registerInputBody:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
  user.registerResponse:
    properties:
      data:
        $ref: '#/definitions/model.User'
      message:
        type: string
    type: object
  user.updateSelfInfoBody:
    properties:
      display_name:
        type: string
      email:
        type: string
    required:
    - email
    type: object
  user.updateSelfInfoResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
  description: API Swagger for Bookmark-Management.
  title: Bookmark Management API
  version: 4.0.0
paths:
  /genpass:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HealthCheckResponse'
      summary: Generate a random password
      tags:
      - password
  /health-check:
    get:
      description: Trả về trạng thái hoạt động hiện tại, tên dịch vụ và instance ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HealthCheckResponse'
        "500":
          description: Lỗi hệ thống nội bộ
          schema:
            additionalProperties:
              type: string
            type: object
      summary: System health-check
      tags:
      - health-check
  /v1/bookmarks:
    get:
      consumes:
      - application/json
      description: Get a list of bookmarks of the current user
      parameters:
      - default: 1
        description: page
        format: int32
        in: query
        name: page
        type: integer
      - default: 20
        description: limit
        format: int32
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.Bookmark'
                type: array
            type: object
      security:
      - BearerAuth: []
      summary: Get a list of bookmarks of the current user
      tags:
      - bookmark
    post:
      consumes:
      - application/json
      description: Create a new bookmark for a bookmark group
      parameters:
      - description: Input required
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/bookmark.createBookmarkInput'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            properties:
              data:
                $ref: '#/definitions/model.Bookmark'
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new bookmark for a bookmark group
      tags:
      - bookmark
  /v1/bookmarks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a bookmark of the current user
      parameters:
      - description: Bookmark ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Message'
      security:
      - BearerAuth: []
      summary: Delete a bookmark of the current user
      tags:
      - bookmark
    put:
      consumes:
      - application/json
      description: Update an existing bookmark of the current user
      parameters:
      - description: Bookmark ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Input required
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/bookmark.updateBookmarkBody'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Message'
      security:
      - BearerAuth: []
      summary: Update an existing bookmark of the current user
      tags:
      - bookmark
  /v1/links/redirect/{code}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Shorten code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
      tags:
      - link
  /v1/links/shorten:
    post:
      consumes:
      - application/json
      description: Generate shorten url based on original url that last upto 7 days
      parameters:
      - description: Input required
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/link.shortenInputBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - link
  /v1/self/info:
    get:
      consumes:
      - application/json
      description: get your current information
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/user.getSelfInfoResponse'
      security:
      - BearerAuth: []
      summary: get your current information
      tags:
      - user
    put:
      consumes:
      - application/json
      description: edit your current information
      parameters:
      - description: Input required
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.updateSelfInfoBody'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/user.updateSelfInfoResponse'
      security:
      - BearerAuth: []
      summary: edit your current information
      tags:
      - user
  /v1/users/login:
    post:
      consumes:
      - application/json
      description: Return a jwt token if the input is correct
      parameters:
      - description: Input required
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.loginInputBody'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/user.loginResponse'
      summary: Return a jwt token if the input is correct
      tags:
      - user
  /v1/users/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided information
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.registerInputBody'
      produces:
      - application/json
      responses:
        "200":
          description: Register an user successfully!
          schema:
            $ref: '#/definitions/user.registerResponse'
      summary: Register a new user
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"